Login and example landing page
So I have uploaded a modified version of the login page and landing page for my library database project. I will show off when we meet how to connect it with WAMP server, as well as setting up the database for collecting user data as it is easier to show remotely but I will try and upload it after.