Add files via upload
Created a rudimentary landing page to be used after login. Displays the user's email as a profile link drop down menu, with options for future page accessibility, including a currently functioning button to logout (the others don't do anything when clicked on). Also implemented a hover over button to create a new portfolio, but this too does not go anywhere as I am unsure where it should direct to.