Database Initialization File, Change to Config
When the contents of this file are dragged into wamp's mysql console, the database is created and initialized with user, portfolio, and stock tables. The latter two tables I was unsure as to what information we wanted to have in those, so I left them to be decided.
I changed the config file to reflect the new name I gave for the database, "TRADEX_DATABASE".