Updated Tables, Added Relationships
Updated the sql initializer file to update the tables for the different fields we discussed in our meeting. Also created relationships between the users and portfolios table, as well as the stocks and portfolios table using a junction table to create a many to many relationship.

Note that the easiest way to update your database is to drop the old one using the command "DROP DATABASE tradex_database" into mysql command line and drag the contents of this file back in. Otherwise, you'll have to update the effected tables manually with separate sql instructions.